+++ packages/lumx-react/src/components/date-picker/DatePickerField.stories.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
-import { DatePickerField } from '@lumx/react';
+import { DatePickerField, DatePickerProps } from '@lumx/react';
 
 export default { title: 'DatePickerField' };
 
 export const simpleDatePickerField = ({ theme }) => {
-    const [value, setValue] = React.useState();
+    const [value, setValue] = React.useState<DatePickerProps['value']>();
 
     return <DatePickerField locale="fr" label="Start date" theme={theme} onChange={setValue} value={value} />;
 };